[TASK-22-23] Make survey outro and quit components
